// src/common/enums/alignment.ts
/**
 * 水平方向对齐方式
 * 
 * @export
 * @enum {number}
 */
export enum HorizontalAlignment {
    Left = 0,
    Center = 1,
    Right = 2,
}


/**
 * 垂直方向对齐方式
 * 
 * @export
 * @enum {number}
 */
export enum VerticalAlignment {
    Top = 0,
    Middle = 1,
    Bottom = 2,
}